Exception handling with ResourceNotFoundException

+++ src/main/java/br/com/supera/game/store/services/OrderService.java
@@ -1,6 +1,7 @@
 package br.com.supera.game.store.services;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -11,6 +12,7 @@ import br.com.supera.game.store.entities.Order;
 import br.com.supera.game.store.entities.OrderItem;
 import br.com.supera.game.store.repositories.OrderItemRepository;
 import br.com.supera.game.store.repositories.OrderRepository;
+import br.com.supera.game.store.services.exceptions.ResourceNotFoundException;
 
 @Service
 public class OrderService {
@@ -40,7 +42,8 @@ public class OrderService {
 	}
 
 	public Order findById(Long id) {
-		return repository.findById(id).get();
+		Optional<Order> opt = repository.findById(id);
+		return opt.orElseThrow(() -> new ResourceNotFoundException(Order.class, id));
 	}
 
 	public Order addItem(Order order, OrderItem item) {
@@ -60,14 +63,23 @@ public class OrderService {
 	}
 
 	public Order removeItem(Order order, OrderItem item) {
-		item.setOrder(order);
+		// @formatter:off
+		item.setOrder(order); // setting the order before test of contains to the possible success on orderItem.equals
+		if (!order.getItems().contains(item)) {
+			throw new ResourceNotFoundException(OrderItem.class,
+					String.format("Order Id %d, product Id %d, item price %.4f",
+							order.getId(),
+							item.getProduct().getId(),
+							item.getPrice().doubleValue()));
+		}
+		// @formatter:on
 		itemRepository.delete(item);
 		order.getItems().remove(item);
 		return order;
 	}
 
 	private OrderItem processItemAddition(Order order, OrderItem item) {
-		int index = order.getItems().indexOf(item); // return -1 if does not exist
+		int index = order.getItems().indexOf(item); // return -1 if does not existing
 		if (index != -1) {
 			OrderItem item2 = order.getItems().get(index);
 			Integer newQuantity = item.getQuantity() + item2.getQuantity();
